requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
CurrencyConverter==0.14.4
discord.py==1.5.1
emoji==0.6.0
idna==2.10
multidict==4.7.6
mysql-connector-python==8.0.22
price-parser==0.3.3
protobuf==3.13.0
six==1.15.0
yarl==1.5.1
